Extract perf logic
